// src/app/models/paralelo.ts
import { Ciclo } from './ciclo';
import { EstudiantesComponent } from '../components/estudiantes/estudiantes.component';
import { Asignatura } from './asignatura';
import { Docente } from './docente';

export class Paralelo {
    id: number;
    nombre: string;
    creteAt: String;
    ciclo: Ciclo;
    estado: boolean;


}